+++ src/Continuous/Prog3/Multiforced/CFrameGraph.java
@@ -132,7 +132,7 @@ class CFrameGraph extends CFrame {
       var1.setColor(Color.white);
       var1.fillRect(0, 390, 600, 190);
       var1.fillRect(0, 0, 450, 20);
-      byte var11 = 60;
+      int var11 = 60;
       double var12 = 1.0472;
       short var14 = 430;
       Color var16 = new Color(0.18F, 0.58F, 0.58F);
